Move user serialization functions from 'helpers' to 'serializers/…' file
