+++ app/serializers/v2/event.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 import { dbAdapter } from '../../models';
-import { userSerializerFunction } from '../../controllers/api/v2/helpers';
+import { userSerializerFunction } from './user';
 
 export async function serializeEvents(events) {
   const [userIdsMapping, postIdsMapping, commentIdsMapping] = await getIntIdsMappings(events);
